seperate function from DetailItem.jsx
