+++ src/components/DetailItem.jsx
@@ -1,16 +1,18 @@
 import React, { useState, useEffect } from "react";
-import { useParams, Link } from "react-router-dom"; // 추가
+import { useParams, Link } from "react-router-dom";
 import { getProductsById } from "../api/api";
 import CommentsList from "./CommentsList";
 import "../css/DetailItem.css";
-import { FormatDate } from "../util/FormatDate";
+import UserInfo from "./UserInfo";
+import TagList from "./TagList";
+import QuestionForm from "./QuestionForm";
+import FavoriteCount from "./FavoriteCount";
 
 const DetailItem = () => {
   const { productId } = useParams();
   const [loading, setLoading] = useState(false);
   const [product, setProduct] = useState(null);
   const [error, setError] = useState(null);
-  const [question, setQuestion] = useState("");
 
   useEffect(() => {
     const fetchProductsById = async () => {
@@ -27,10 +29,6 @@ const DetailItem = () => {
 
     fetchProductsById();
   }, [productId]);
-
-  const handleChange = (e) => {
-    setQuestion(e.target.value);
-  };
 
   if (loading) {
     return <p>Loading...</p>;
@@ -69,56 +67,17 @@ const DetailItem = () => {
           <p className="deatilItem-introduce">상품 소개</p>
           <p className="detailItem-description">{product.description}</p>
           <p className="deatilItem-introduce">상품 태그</p>
-          <div className="detailItem-tag-container">
-            {product.tags.map((tag, index) => (
-              <div key={index} className="detailItem-tag-box">
-                #{tag}
-              </div>
-            ))}
-          </div>
+          <TagList tags={product.tags} />
           <div className="user-container">
-            <div className="user-box">
-              <img
-                className="user-profile"
-                src="/assets/size=large.png"
-                alt="사용자 프로필 이미지"
-              />
-              <div>
-                <p className="user-nickname">{product.ownerNickname}</p>
-                <p className="created-time">{FormatDate(product.createdAt)}</p>
-              </div>
-            </div>
-            <div className="favoriteCount-container">
-              <div className="favoriteCount-box">
-                <img
-                  className="favoriteCount-image"
-                  src="/assets/Icon (2).png"
-                  alt="좋아요 하트 모양"
-                />
-                <p className="favoriteCount">{product.favoriteCount}</p>
-              </div>
-            </div>
+            <UserInfo
+              ownerNickname={product.ownerNickname}
+              createdAt={product.createdAt}
+            />
+            <FavoriteCount count={product.favoriteCount} />
           </div>
         </div>
       </section>
-      <section className="question-container">
-        <p className="question-header">문의하기</p>
-        <form>
-          <label htmlFor="question-input"></label>
-          <textarea
-            placeholder="개인정보를 공유 및 요청하거나, 명예 훼손, 무단 광고, 불법 정보 유포시 모니터링 후 삭제될 수 있으며, 이에 대한 민형사상 책임은 게시자에게 있습니다."
-            id="question-input"
-            type="text"
-            name="question"
-            onChange={handleChange}
-          ></textarea>
-          <div className="button-box">
-            <button className="question-register-button" disabled={!question}>
-              등록
-            </button>
-          </div>
-        </form>
-      </section>
+      <QuestionForm />
       <CommentsList />
       <Link to="/items" className="back-button-link">
         <button className="back-button">
